Narrow mdlFile 00342130 pointer parameter

+++ src/Event/Fcl/fclShopBGManager.c
@@ -47,7 +47,7 @@ u32 FUN_00411790(u64 param_1,int param_2,long param_3);
 u64 FUN_00411ee0(u64 param_1,int param_2);
 u32 FUN_004120a0(u64 param_1, int param_2, int param_3);
 u64 FUN_00412580(u64 param_1,int param_2);
-u32 FUN_00412610(u64 param_1, int param_2, int param_3);
+u32 FUN_00412610(u64 param_1, int param_2, long param_3);
 u64 FUN_00412af0(u64 param_1,int param_2);
 u32 FUN_00412bc0(u64 param_1,int param_2,long param_3);
 u64 FUN_00413010(u64 param_1,int param_2);
@@ -2176,7 +2176,7 @@ u64 FUN_00412580(u64 param_1,int param_2)
 // FUN_00412610 NONMATCHING
 
 
-u32 FUN_00412610(u64 param_1, int param_2, int param_3) { short sVar1;
+u32 FUN_00412610(u64 param_1, int param_2, long param_3) { short sVar1;
 
 int iVar2;
